// apps/admin/src/api/generated/model/assistantThreadDtoOutputTurnsItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */
import type { AssistantThreadDtoOutputTurnsItemAssistant } from "./assistantThreadDtoOutputTurnsItemAssistant";
import type { AssistantThreadDtoOutputTurnsItemEffort } from "./assistantThreadDtoOutputTurnsItemEffort";
import type { AssistantThreadDtoOutputTurnsItemError } from "./assistantThreadDtoOutputTurnsItemError";
import type { AssistantThreadDtoOutputTurnsItemModel } from "./assistantThreadDtoOutputTurnsItemModel";
import type { AssistantThreadDtoOutputTurnsItemStatus } from "./assistantThreadDtoOutputTurnsItemStatus";
import type { AssistantThreadDtoOutputTurnsItemUser } from "./assistantThreadDtoOutputTurnsItemUser";

export type AssistantThreadDtoOutputTurnsItem = {
  /** @nullable */
  assistant: AssistantThreadDtoOutputTurnsItemAssistant;
  /**
   * @maximum 9007199254740991
   * @exclusiveMinimum 0
   */
  attempt: number;
  /**
   * @nullable
   * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
   */
  completedAt: string | null;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  createdAt: string;
  effort: AssistantThreadDtoOutputTurnsItemEffort;
  /** @nullable */
  error: AssistantThreadDtoOutputTurnsItemError;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  model: AssistantThreadDtoOutputTurnsItemModel;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  requestId: string;
  /**
   * @maximum 9007199254740991
   * @exclusiveMinimum 0
   */
  sequence: number;
  /**
   * @nullable
   * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
   */
  startedAt: string | null;
  status: AssistantThreadDtoOutputTurnsItemStatus;
  user: AssistantThreadDtoOutputTurnsItemUser;
};